legend-star/backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.24.0
motor==4.4.0
python-dotenv==1.0.1
pydantic==2.7.0
